web3js: Separate each usecases
