+++ web3js-ext/example/accountKey/update_AccountKeyPublic.js
@@ -1,0 +1,36 @@
+// AccountKeyPublic
+// https://docs.klaytn.foundation/docs/learn/accounts/
+
+const { Web3, TxType, AccountKeyType, toPeb, getPublicKeyFromPrivate } = require("@klaytn/web3js-ext");
+
+// Using senderPriv == senderNewPriv to execute this example repeatedly.
+// But you might want to register a different private key.
+const senderAddr = "0xfb60ded0ae96fe04eed6450aead860aa9d57128e";
+const senderPriv = "0x0e4ca6d38096ad99324de0dde108587e5d7c600165ae4cd6c2462c597458c2b8";
+const senderNewPriv = "0x0e4ca6d38096ad99324de0dde108587e5d7c600165ae4cd6c2462c597458c2b8";
+
+const provider = new Web3.providers.HttpProvider("https://public-en-baobab.klaytn.net");
+const web3 = new Web3(provider);
+const senderAccount = web3.eth.accounts.privateKeyToAccount(senderPriv);
+
+async function main() {
+  const senderNewPub = getPublicKeyFromPrivate(senderNewPriv);
+  console.log({ senderNewPub });
+
+  const tx = {
+    type: TxType.AccountUpdate,
+    from: senderAddr,
+    key: {
+      type: AccountKeyType.Public,
+      key: senderNewPub,
+    }
+  };
+
+  const signResult = await senderAccount.signTransaction(tx);
+  console.log("signedTx", signResult.transactionHash);
+
+  const receipt = await web3.eth.sendSignedTransaction(signResult.rawTransaction);
+  console.log("receipt", receipt);
+}
+
+main().catch(console.error);
