+++ web3js-ext/example/accountKey/update_AccountKeyRoleBased.js
@@ -1,20 +1,18 @@
 // AccountKeyRoleBased
 // https://docs.klaytn.foundation/docs/learn/accounts/
 
-const { Web3, TxType, AccountKeyType, toPeb, getPublicKeyFromPrivate } = require("@klaytn/web3js-ext");
+const { Web3, TxType, AccountKeyType, getPublicKeyFromPrivate } = require("@klaytn/web3js-ext");
 
 const senderAddr = "0x334b4d3c775c45c59de54e9f0408cba25a1aece7";
 const senderRoleTransactionPriv = "0xc9668ccd35fc20587aa37a48838b48ccc13cf14dd74c8999dd6a480212d5f7ac";
 const senderRoleAccountUpdatePriv = "0x9ba8cb8f60044058a9e6f815c5c42d3a216f47044c61a1750b6d29ddc7f34bda";
 const senderRoleFeePayerPriv = "0x0e4ca6d38096ad99324de0dde108587e5d7c600165ae4cd6c2462c597458c2b8";
-const receiverAddr = "0xc40b6909eb7085590e1c26cb3becc25368e249e9";
 
 const provider = new Web3.providers.HttpProvider("https://public-en-baobab.klaytn.net");
 const web3 = new Web3(provider);
 const updaterAccount = web3.eth.accounts.privateKeyToAccount(senderRoleAccountUpdatePriv);
-const txAccount = web3.eth.accounts.privateKeyToAccount(senderRoleTransactionPriv);
 
-async function updateAccount() {
+async function main() {
   const pub1 = getPublicKeyFromPrivate(senderRoleTransactionPriv);
   const pub2 = getPublicKeyFromPrivate(senderRoleAccountUpdatePriv);
   const pub3 = getPublicKeyFromPrivate(senderRoleFeePayerPriv);
@@ -41,40 +39,4 @@ async function updateAccount() {
   console.log("receipt", receipt);
 }
 
-async function sendTx() {
-  let tx = {
-    type: TxType.ValueTransfer,
-    from: senderAddr,
-    to: receiverAddr,
-    value: toPeb("0.01", "KLAY"),
-    gasLimit: 100000,
-  };
-
-  const signResult = await txAccount.signTransaction(tx);
-  console.log("signedTx", signResult.transactionHash);
-
-  const receipt = await web3.eth.sendSignedTransaction(signResult.rawTransaction);
-  console.log("receipt", receipt);
-}
-
-async function recoverMsg() {
-  const msg = "hello";
-  const msghex = Web3.utils.utf8ToHex(msg);
-  const signResult = txAccount.sign(msg);
-  console.log({ senderAddr, msg, msghex, sig: signResult.signature });
-
-  const { v, r, s } = signResult;
-  const addr1 = web3.eth.accounts.recover(msg, v, r, s);
-  console.log("recoveredAddr lib", addr1, addr1.toLowerCase() === txAccount.address.toLowerCase());
-
-  const sig = signResult.signature;
-  const addr2 = await web3.klay.recoverFromMessage(senderAddr, msghex, sig, "latest");
-  console.log("recoveredAddr rpc", addr2, addr2.toLowerCase() === txAccount.address.toLowerCase());
-}
-
-async function main() {
-  await updateAccount();
-  await sendTx();
-  await recoverMsg();
-}
 main().catch(console.error);
